backend_momono/momono_hizkia/requirements.txt
pyramid==2.0
pyramid_debugtoolbar==4.12.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
waitress==2.1.2
alembic==1.12.1
pyjwt==2.8.0
bcrypt==4.0.1
pytest==7.4.3
webtest==3.0.1
zope.sqlalchemy==1.6
